requirements.txt
addict
torchvision
torch-fidelity
tqdm
pillow
numpy
opencv-python
yapf
matplotlib
